First yoke design to be 3D printed.
This will let me get the full measurements right for the 3D model, for medium to large hand sizes.

Thumbs will be annoying to get right initially.

Going to keep this one simple with no switches yet as I need to design the PCB to control the force feedback tensioner.